+++ src/main/java/com/zhuanche/entity/driver/FinancialClueGoods.java
@@ -38,7 +38,7 @@ public class FinancialClueGoods {
 
     private BigDecimal guidancePrice;
 
-    private Integer discharge;
+    private Double discharge;
 
     private Integer mileage;
 
@@ -52,9 +52,9 @@ public class FinancialClueGoods {
 
     private String environmentalProtectionStandard;
 
-    private Integer fastChargingTime;
+    private Double fastChargingTime;
 
-    private Integer slowChargingTime;
+    private Double slowChargingTime;
 
     private Double fastPercentage;
 
@@ -224,11 +224,11 @@ public class FinancialClueGoods {
         this.guidancePrice = guidancePrice;
     }
 
-    public Integer getDischarge() {
+    public Double getDischarge() {
         return discharge;
     }
 
-    public void setDischarge(Integer discharge) {
+    public void setDischarge(Double discharge) {
         this.discharge = discharge;
     }
 
@@ -280,19 +280,19 @@ public class FinancialClueGoods {
         this.environmentalProtectionStandard = environmentalProtectionStandard == null ? null : environmentalProtectionStandard.trim();
     }
 
-    public Integer getFastChargingTime() {
+    public Double getFastChargingTime() {
         return fastChargingTime;
     }
 
-    public void setFastChargingTime(Integer fastChargingTime) {
+    public void setFastChargingTime(Double fastChargingTime) {
         this.fastChargingTime = fastChargingTime;
     }
 
-    public Integer getSlowChargingTime() {
+    public Double getSlowChargingTime() {
         return slowChargingTime;
     }
 
-    public void setSlowChargingTime(Integer slowChargingTime) {
+    public void setSlowChargingTime(Double slowChargingTime) {
         this.slowChargingTime = slowChargingTime;
     }
